deleted:    __pycache__/visualFeedback.cpython-37.pyc 	deleted:    archive/GUI.py 	deleted:    archive/GUI_2.py 	deleted:    archive/GUI_visualFeedback.py 	deleted:    archive/data_stream.csv 	deleted:    archive/eoj_ekg.csv 	deleted:    archive/fakeStream.py 	deleted:    archive/group_17_ekg_test_1.xlsx 	deleted:    archive/kelvinToRGB.py 	deleted:    archive/matOut.csv 	deleted:    archive/mindfulness_monitor.py 	deleted:    archive/mindfulness_states.py 	deleted:    archive/readADC.py 	deleted:    archive/screen_init.py 	deleted:    archive/visual feedback.py 	deleted:    archive/visualFeedback.py 	deleted:    archive/visual_feedback.py 	deleted:    mindfulnessMonitor.py 	renamed:    Util.py -> src/Util/Util.py 	new file:   src/Util/__init__.py 	renamed:    signalProcess.py -> src/Util/signalProcess.py 	renamed:    hapticFeedback.py -> src/hapticFeedback.py 	new file:   src/mindfulnessMonitor.py 	renamed:    visualFeedback.py -> src/visualFeedback.py
